fix: Support SSSS 16 correctly in lj92.
As explained [here](https://www.w3.org/Graphics/JPEG/itu-t81.pdf) in
section H.1.2.2 on page 134, the correct number of extra bits for SSSS
16 is 0 because the difference value is always 32768.  The lj92 code
was assuming that SSSS is always equal to the number of extra bits and
then doing some math using the pattern that appears in the first 16
elements, extending that pattern to SSSS 16.  That is the wrong thing
to do.

The solution is to recognize that the number of extra bits is not
necessarily the same as the SSSS and then put in a special case for
SSSS 16.